Conflicts:
	src/org/broad/igv/PreferenceManager.java
	src/org/broad/igv/batch/CommandExecutor.java
	src/org/broad/igv/cbio/FilterGeneNetworkUI.java
	src/org/broad/igv/cbio/FilterGeneNetworkUI.jfd
	src/org/broad/igv/dev/IParser.java
	src/org/broad/igv/dev/SegmentedReader.java
	src/org/broad/igv/dev/db/DBManager.java
	src/org/broad/igv/dev/db/SQLLineParserByName.java
	src/org/broad/igv/feature/Chromosome.java
	src/org/broad/igv/feature/FeatureUtils.java
	src/org/broad/igv/feature/genome/GenomeManager.java
	src/org/broad/igv/hic/HeatmapPanel.java
	src/org/broad/igv/hic/HeatmapRenderer.java
	src/org/broad/igv/hic/HiC.java
	src/org/broad/igv/hic/MainWindow.java
	src/org/broad/igv/hic/data/Dataset.java
	src/org/broad/igv/hic/data/DatasetReader.java
	src/org/broad/igv/hic/data/DensityFunction.java
	src/org/broad/igv/hic/data/MatrixZoomData.java
	src/org/broad/igv/hic/matrix/DiskResidentBlockMatrix.java
	src/org/broad/igv/hic/matrix/DiskResidentRowMatrix.java
	src/org/broad/igv/hic/tools/DensityCalculation.java
	src/org/broad/igv/hic/tools/DensityUtil.java
	src/org/broad/igv/hic/tools/FragmentCalculation.java
	src/org/broad/igv/hic/tools/HiCTools.java
	src/org/broad/igv/hic/tools/Preprocessor.java
	src/org/broad/igv/hic/track/EigenvectorTrack.java
	src/org/broad/igv/hic/track/TrackPanel.java
	src/org/broad/igv/renderer/IGVFeatureRenderer.java
	src/org/broad/igv/sam/AlignmentDataManager.java
	src/org/broad/igv/sam/AlignmentInterval.java
	src/org/broad/igv/sam/AlignmentRenderer.java
	src/org/broad/igv/sam/AlignmentTrack.java
	src/org/broad/igv/sam/DenseAlignmentCounts.java
	src/org/broad/igv/track/AnalysisDialog.java
	src/org/broad/igv/track/CombinedFeatureSource.java
	src/org/broad/igv/track/DataTrack.java
	src/org/broad/igv/track/FeatureTrack.java
	src/org/broad/igv/track/PackedFeatures.java
	src/org/broad/igv/track/TrackMenuUtils.java
	src/org/broad/igv/ui/GenomeSelectionDialog.java
	src/org/broad/igv/ui/GenomeSelectionDialog.jfd
	src/org/broad/igv/ui/IGV.java
	src/org/broad/igv/ui/IGVCommandBar.java
	src/org/broad/igv/ui/Main.java
	src/org/broad/igv/ui/PreferencesEditor.java
	src/org/broad/igv/ui/PreferencesEditor.jfd
	src/org/broad/igv/ui/ResourceTree.java
	src/org/broad/igv/ui/action/LoadFromServerAction.java
	src/org/broad/igv/ui/panel/AttributeHeaderPanel.java
	src/org/broad/igv/ui/panel/ReferenceFrame.java
	src/org/broad/igv/ui/util/MessageUtils.java
	src/org/broad/igv/util/FileUtils.java
	src/org/broad/igv/util/HttpUtils.java
	src/org/broad/igv/util/Utilities.java
	src/org/broad/igv/util/collections/CachedIntervals.java
	src/org/broad/igv/util/collections/CollUtils.java
	test/src/org/broad/igv/ui/MainTest.java

+++ src/org/iontorrent/sam2flowgram/util/AlignUtil.java
@@ -333,7 +333,7 @@ public class AlignUtil {
      * @param flowSignal the 100 scaled flow signal
      * @return the base call represented by this flow signal
      */
-    public static int getBaseCallFromFlowSignal(int flowSignal) {
+    public static int getBaseCallFromFlowSignal(double flowSignal) {
         //System.err.println(flowSignal + " => " + (int)((flowSignal + 50.0) / 100.0));
         return (int) ((flowSignal + 50.0) / 100.0);
     }
